+++ app/src/main/java/com/prof18/secureqrreader/screens/ResultScreen.kt
@@ -1,4 +1,20 @@
-package com.prof18.secureqrreader
+/*
+ * Copyright 2022 Marco Gomiero
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.prof18.secureqrreader.screens
 
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Arrangement
@@ -49,7 +65,7 @@ fun ResultScreen(
                     Text(
                         modifier = Modifier
                             .fillMaxWidth(),
-                        text = "Sorry, something unexpected happened. Please retry", // TODO: localise
+                        text = stringResource(string.unespected_error),
                         style = MaterialTheme.typography.body1,
                         color = MaterialTheme.colors.onBackground,
                     )
