+++ app/src/main/java/com/prof18/secureqrreader/screens/WelcomeScreen.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.prof18.secureqrreader
+package com.prof18.secureqrreader.screens
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Image
@@ -35,6 +35,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
+import com.prof18.secureqrreader.R.drawable
+import com.prof18.secureqrreader.R.string
 import com.prof18.secureqrreader.style.Margins
 import com.prof18.secureqrreader.style.SecureQrReaderTheme
 
@@ -42,49 +44,47 @@ import com.prof18.secureqrreader.style.SecureQrReaderTheme
 internal fun WelcomeScreen(
     onStartClick: () -> Unit = {},
 ) {
-    SecureQrReaderTheme {
-        Column(
-            modifier = Modifier.fillMaxSize(),
+    Column(
+        modifier = Modifier.fillMaxSize(),
+    ) {
+        LazyColumn(
+            modifier = Modifier.weight(1f)
         ) {
-            LazyColumn(
-                modifier = Modifier.weight(1f)
-            ) {
-                item {
-                    Text(
-                        modifier = Modifier
-                            .padding(top = Margins.big)
-                            .fillMaxWidth()
-                            .wrapContentWidth(align = Alignment.CenterHorizontally),
-                        color = MaterialTheme.colors.onBackground,
-                        text = stringResource(id = R.string.app_name),
-                        style = MaterialTheme.typography.h1.copy(fontSize = 28.sp),
-                    )
-                }
-                item {
-                    Text(
-                        modifier = Modifier
-                            .padding(horizontal = Margins.big)
-                            .padding(top = Margins.big),
-                        color = MaterialTheme.colors.onBackground,
-                        text = stringResource(id = R.string.welcome_screen_content),
-                        style = MaterialTheme.typography.body1,
-                    )
-                }
-                item {
-                    Image(
-                        painter = painterResource(id = R.drawable.privacy_vector),
-                        contentDescription = null,
-                    )
-                }
+            item {
+                Text(
+                    modifier = Modifier
+                        .padding(top = Margins.big)
+                        .fillMaxWidth()
+                        .wrapContentWidth(align = Alignment.CenterHorizontally),
+                    color = MaterialTheme.colors.onBackground,
+                    text = stringResource(id = string.app_name),
+                    style = MaterialTheme.typography.h1.copy(fontSize = 28.sp),
+                )
             }
-            Button(
-                modifier = Modifier
-                    .align(Alignment.End)
-                    .padding(Margins.regular),
-                onClick = onStartClick
-            ) {
-                Text(stringResource(id = R.string.welcome_screen_button))
+            item {
+                Text(
+                    modifier = Modifier
+                        .padding(horizontal = Margins.big)
+                        .padding(top = Margins.big),
+                    color = MaterialTheme.colors.onBackground,
+                    text = stringResource(id = string.welcome_screen_content),
+                    style = MaterialTheme.typography.body1,
+                )
             }
+            item {
+                Image(
+                    painter = painterResource(id = drawable.privacy_vector),
+                    contentDescription = null,
+                )
+            }
+        }
+        Button(
+            modifier = Modifier
+                .align(Alignment.End)
+                .padding(Margins.regular),
+            onClick = onStartClick
+        ) {
+            Text(stringResource(id = string.welcome_screen_button))
         }
     }
 }
